requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
streamlit==1.27.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
face-recognition==1.3.0
dlib==19.24.2
mediapipe==0.10.3
